moving cli modules out of src root
