+++ framework/bin/generate-benchmark-from-open-api.js
@@ -1,4 +1,4 @@
 #!/usr/bin/env node
-import create from '../src/open-api-route-builder.js';
+import create from '../src/cli/open-api-route-builder.ts';
 
 create(process.argv, process.cwd(),);
