+++ framework/bin/make-benchmark-project.js
@@ -1,4 +1,4 @@
 #!/usr/bin/env node
-import make from '../src/make.js';
+import make from '../src/cli/make.ts';
 
 make(process.argv, process.cwd(),);
