+++ framework/bin/run-benchmark.js
@@ -1,4 +1,4 @@
 #!/usr/bin/env node
-import cli from '../src/cli.js';
+import cli from '../src/cli/cli.ts';
 
 cli(process.argv,);
