+++ framework/bin/run-loadtest.js
@@ -1,4 +1,4 @@
 #!/usr/bin/env node
-import load from '../src/loader.js';
+import load from '../src/cli/loader.ts';
 
 load(process.argv,);
